ローカル通知のinitial処理をmain.dartからRepositoryに移動
